requirements.txt
beautifulsoup4==4.12.2
Django==4.2.4
feedparser==6.0.11
gnews==0.3.8
python_dateutil==2.8.2
Requests==2.31.0
scikit_learn==1.3.0
sentence_transformers==3.2.1
transformers==4.45.2
